vrp_project/vrp_project/requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
idna==2.10
importlib-metadata==2.0.0
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
oauthlib==3.1.0
openapi-codec==1.3.2
ortools==7.8.7959
prompt-toolkit==3.0.7
protobuf==3.13.0
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
zipp==3.3.0
